add bitgo auth utils for bitgo express
